# back-end/music-selector/src/recommendations/recommendations.controller.ts
import {
  Controller,
  Post,
  Get,
  Body,
  UseGuards,
  Request,
  Param,
  Query,
  Logger,
  Delete,
} from '@nestjs/common';
import { ApiTags, ApiOperation, ApiResponse, ApiBearerAuth, ApiParam, ApiQuery } from '@nestjs/swagger';
import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
import { RecommendationsService } from './recommendations.service';
import { GetRecommendationsDto } from './dto/get-recommendations.dto';
import { SaveFeedbackDto } from './dto/save-feedback.dto';

@Controller('api/recommendations')
@UseGuards(JwtAuthGuard)
@ApiTags('Recommendations')
@ApiBearerAuth()
export class RecommendationsController {
  private readonly logger = new Logger(RecommendationsController.name);

  constructor(private readonly recommendationsService: RecommendationsService) {}

  /**
   * POST /api/recommendations/generate
   * RN17-RN22: Gerar recomendações sob demanda
   */
  @Post('generate')
  @ApiOperation({
    summary: 'Gerar recomendações sob demanda',
    description: 'RN17-RN22: Cria vibe customizada com Objetivo + Energia + Humor. Retorna 10 faixas ordenadas por relevância',
  })
  @ApiResponse({
    status: 200,
    description: 'Recomendações geradas com sucesso',
    schema: {
      example: {
        playlistId: 'uuid',
        playlistName: 'My Vibe',
        objective: 'FOCUS',
        mood: 'HAPPY',
        energyLevel: 'HIGH',
        generatedAt: '2026-05-14T12:00:00Z',
        tracks: [
          {
            id: 'spotify_id',
            title: 'Song Name',
            artist: 'Artist Name',
            features: { energy: 0.75, valence: 0.65, danceability: 0.7 },
          },
        ],
        totalTracks: 10,
      },
    },
  })
  @ApiResponse({ status: 401, description: 'Token inválido ou expirado' })
  async getRecommendations(
    @Request() req,
    @Body() dto: GetRecommendationsDto,
  ) {
    this.logger.log(
      `Gerando recomendações para usuário ${req.user.id} com objetivo: ${dto.objective}`,
    );
    return this.recommendationsService.getRecommendations(req.user.id, dto);
  }

  /**
   * GET /api/recommendations/daily-vibe
   * RN10, RN14-RN15: Gerar Vibe Diária automática
   */
  @Get('daily-vibe')
  @ApiOperation({
    summary: 'Gerar Vibe Diária',
    description: 'RN10, RN14-RN15: Recomendação automática recalculada a cada 24h baseada no perfil do usuário',
  })
  @ApiResponse({
    status: 200,
    description: 'Vibe diária gerada com sucesso',
  })
  @ApiResponse({ status: 401, description: 'Não autenticado' })
  async getDailyVibe(@Request() req) {
    this.logger.log(`Gerando vibe diária para usuário ${req.user.id}`);
    return this.recommendationsService.generateDailyVibe(req.user.id);
  }

  /**
   * GET /api/recommendations/vibes
   * RN14-RN15: Recuperar Vibes Diárias (playlists automáticas)
   */
  @Get('vibes')
  @ApiOperation({
    summary: 'Listar Vibes Diárias',
    description: 'RN14-RN15: Retorna histórico de vibes diárias do usuário',
  })
  @ApiResponse({
    status: 200,
    description: 'Lista de vibes diárias',
    isArray: true,
  })
  async getUserDailyVibes(@Request() req) {
    this.logger.log(`Recuperando vibes diárias do usuário ${req.user.id}`);
    return this.recommendationsService.getUserDailyVibes(req.user.id);
  }

  /**
   * GET /api/recommendations/history
   * RN14: Recuperar histórico de playlists
   */
  @Get('history')
  @ApiOperation({
    summary: 'Histórico de playlists',
    description: 'RN14: Retorna lista paginada de playlists criadas pelo usuário',
  })
  @ApiQuery({
    name: 'limit',
    required: false,
    type: Number,
    description: 'Limite de resultados (padrão: 10)',
  })
  @ApiResponse({
    status: 200,
    description: 'Histórico de playlists',
    isArray: true,
  })
  async getPlaylistHistory(
    @Request() req,
    @Query('limit') limit?: number,
  ) {
    this.logger.log(`Recuperando histórico de playlists do usuário ${req.user.id}`);
    return this.recommendationsService.getUserPlaylistHistory(req.user.id, limit);
  }

  /**
   * GET /api/recommendations/feedback
   * Recuperar histórico de feedback do usuário
   */
  @Get('feedback')
  @ApiOperation({
    summary: 'Histórico de feedback',
    description: 'RN23-RN24: Retorna histórico de likes/dislikes do usuário',
  })
  @ApiQuery({
    name: 'limit',
    required: false,
    type: Number,
    description: 'Limite de resultados (padrão: 50)',
  })
  @ApiResponse({
    status: 200,
    description: 'Histórico de feedback',
    isArray: true,
  })
  async getUserFeedback(
    @Request() req,
    @Query('limit') limit?: number,
  ) {
    this.logger.log(`Recuperando histórico de feedback do usuário ${req.user.id}`);
    return this.recommendationsService.getUserFeedbackHistory(req.user.id, limit);
  }

  /**
   * GET /api/recommendations/feedback/stats
   * Recuperar estatísticas de feedback
   */
  @Get('feedback/stats')
  @ApiOperation({
    summary: 'Estatísticas de feedback',
    description: 'Retorna análise de padrões de likes/dislikes por contexto',
  })
  @ApiResponse({
    status: 200,
    description: 'Estatísticas de feedback',
    schema: {
      example: {
        totalLikes: 42,
        totalDislikes: 8,
        likesByObjective: { FOCUS: 20, RELAX: 15, WORKOUT: 7 },
      },
    },
  })
  async getUserFeedbackStats(@Request() req) {
    this.logger.log(`Recuperando estatísticas de feedback do usuário ${req.user.id}`);
    return this.recommendationsService.getUserFeedbackStats(req.user.id);
  }

  /**
   * POST /api/recommendations/feedback
   * RN23-RN24: Registrar feedback (like/dislike)
   */
  @Post('feedback')
  @ApiOperation({
    summary: 'Registrar feedback de música',
    description: 'RN23-RN24: Registra like/dislike e bloqueia música em contexto (dislike). Melhora futuras recomendações',
  })
  @ApiResponse({
    status: 201,
    description: 'Feedback registrado com sucesso',
    schema: {
      example: {
        id: 'feedback_uuid',
        userId: 'user_uuid',
        trackId: 'track_uuid',
        reaction: 'LIKE',
        objectiveContext: 'FOCUS',
        createdAt: '2026-05-14T12:00:00Z',
      },
    },
  })
  @ApiResponse({ status: 400, description: 'Dados de feedback inválidos' })
  async recordFeedback(
    @Request() req,
    @Body() dto: SaveFeedbackDto,
  ) {
    this.logger.log(
      `Registrando ${dto.reaction} do usuário ${req.user.id} para track ${dto.trackId}`,
    );
    return this.recommendationsService.recordFeedback(
      req.user.id,
      dto.trackId,
      dto.reaction,
      dto.objectiveContext,
    );
  }

  /**
   * GET /api/recommendations/:playlistId
   * Recuperar detalhes de uma playlist específica
   */
  @Get(':playlistId')
  @ApiParam({
    name: 'playlistId',
    description: 'ID da playlist',
  })
  @ApiOperation({
    summary: 'Detalhes da playlist',
    description: 'Retorna informações completas de uma playlist incluindo todas as faixas',
  })
  @ApiResponse({
    status: 200,
    description: 'Detalhes da playlist',
  })
  @ApiResponse({ status: 404, description: 'Playlist não encontrada' })
  async getPlaylistDetails(
    @Request() req,
    @Param('playlistId') playlistId: string,
  ) {
    this.logger.log(`Recuperando detalhes da playlist ${playlistId}`);
    return this.recommendationsService.getPlaylistDetails(playlistId, req.user.id);
  }

  /**
   * DELETE /api/recommendations/:playlistId
   * Deletar uma playlist específica
   */
  @Delete(':playlistId')
  @ApiParam({
    name: 'playlistId',
    description: 'ID da playlist',
  })
  @ApiOperation({
    summary: 'Deletar playlist',
    description: 'Remove uma playlist criada pelo usuário',
  })
  @ApiResponse({
    status: 200,
    description: 'Playlist deletada com sucesso',
  })
  @ApiResponse({ status: 404, description: 'Playlist não encontrada' })
  @ApiResponse({ status: 403, description: 'Não autorizado a deletar esta playlist' })
  async deletePlaylist(
    @Request() req,
    @Param('playlistId') playlistId: string,
  ) {
    this.logger.log(`Deletando playlist ${playlistId} do usuário ${req.user.id}`);
    return this.recommendationsService.deletePlaylist(playlistId, req.user.id);
  }
}