# back-end/music-selector/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MusicSelector API",
    "description": "API de recomendação musical com IA - Sistema multiplataforma de recomendação musical com classificação por estado emocional e contexto situacional",
    "version": "1.0.0",
    "contact": {
      "name": "MusicSelector Team",
      "url": "https://musicSelector.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RecommendationRequest": {
        "type": "object",
        "required": ["objective", "mood", "energyLevel"],
        "properties": {
          "objective": {
            "type": "string",
            "enum": ["FOCUS", "RELAX", "WORKOUT", "PARTY"],
            "example": "FOCUS"
          },
          "mood": {
            "type": "string",
            "enum": ["HAPPY", "NEUTRAL", "ANXIOUS", "SAD"],
            "example": "HAPPY"
          },
          "energyLevel": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH"],
            "example": "HIGH"
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50,
            "example": 10
          }
        }
      },
      "Track": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "album": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "popularity": {
            "type": "integer"
          },
          "features": {
            "type": "object",
            "properties": {
              "energy": {
                "type": "number"
              },
              "valence": {
                "type": "number"
              },
              "danceability": {
                "type": "number"
              }
            }
          }
        }
      },
      "RecommendationResponse": {
        "type": "object",
        "properties": {
          "playlistId": {
            "type": "string"
          },
          "playlistName": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "mood": {
            "type": "string"
          },
          "energyLevel": {
            "type": "string"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Track"
            }
          },
          "totalTracks": {
            "type": "integer"
          }
        }
      },
      "FeedbackRequest": {
        "type": "object",
        "required": ["trackId", "reaction", "objectiveContext"],
        "properties": {
          "trackId": {
            "type": "string",
            "example": "spotify_track_id"
          },
          "reaction": {
            "type": "string",
            "enum": ["LIKE", "DISLIKE"],
            "example": "LIKE"
          },
          "objectiveContext": {
            "type": "string",
            "enum": ["FOCUS", "RELAX", "WORKOUT", "PARTY"],
            "example": "FOCUS"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/recommendations/generate": {
      "post": {
        "tags": ["Recommendations"],
        "summary": "Gerar recomendações sob demanda",
        "description": "RN17-RN22: Cria vibe customizada com Objetivo + Energia + Humor. Retorna 10 faixas ordenadas por relevância",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recomendações geradas com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido ou expirado"
          }
        }
      }
    },
    "/api/recommendations/daily-vibe": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Gerar Vibe Diária",
        "description": "RN10, RN14-RN15: Recomendação automática recalculada a cada 24h baseada no perfil do usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Vibe diária gerada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado"
          }
        }
      }
    },
    "/api/recommendations/vibes": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Listar Vibes Diárias",
        "description": "RN14-RN15: Retorna histórico de vibes diárias do usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de vibes diárias",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecommendationResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/recommendations/history": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Histórico de playlists",
        "description": "RN14: Retorna lista paginada de playlists criadas pelo usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Limite de resultados (padrão: 10)",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico de playlists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecommendationResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/recommendations/feedback": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Histórico de feedback",
        "description": "RN23-RN24: Retorna histórico de likes/dislikes do usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Limite de resultados (padrão: 50)",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico de feedback",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Recommendations"],
        "summary": "Registrar feedback de música",
        "description": "RN23-RN24: Registra like/dislike e bloqueia música em contexto (dislike). Melhora futuras recomendações",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback registrado com sucesso"
          },
          "400": {
            "description": "Dados de feedback inválidos"
          }
        }
      }
    },
    "/api/recommendations/feedback/stats": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Estatísticas de feedback",
        "description": "Retorna análise de padrões de likes/dislikes por contexto",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Estatísticas de feedback",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalLikes": {
                      "type": "integer"
                    },
                    "totalDislikes": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/recommendations/{playlistId}": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Detalhes da playlist",
        "description": "Retorna informações completas de uma playlist incluindo todas as faixas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "playlistId",
            "in": "path",
            "required": true,
            "description": "ID da playlist",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes da playlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Playlist não encontrada"
          }
        }
      },
      "delete": {
        "tags": ["Recommendations"],
        "summary": "Deletar playlist",
        "description": "Remove uma playlist criada pelo usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "playlistId",
            "in": "path",
            "required": true,
            "description": "ID da playlist",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist deletada com sucesso"
          },
          "404": {
            "description": "Playlist não encontrada"
          },
          "403": {
            "description": "Não autorizado a deletar esta playlist"
          }
        }
      }
    }
  },
...
}